Move setIn to utils
